Changed How Definition Closures Work
Currently, attributes always override definitions. There are cases where I don't want it to be overridden, but changed based on the input. To achieve this, I changed how closures work in fixture definitions. Instead of getting an instance and the attributes, it will get the value from the attribute and it can return the value it needs.

Remove the lazy method because that used the old way callbacks worked.

Allow states to return just an array for simple states.